:bug: downloading now works

+++ src/main/kotlin/Main.kt
@@ -10,12 +10,16 @@ import com.github.ajalt.mordant.rendering.Whitespace
 import com.github.ajalt.mordant.terminal.Terminal
 import commands.about
 import commands.help
+import okhttp3.OkHttpClient
+import okhttp3.Request
 import java.net.InetAddress
 import kotlin.system.exitProcess
 import java.lang.ProcessBuilder.Redirect
 import java.util.concurrent.TimeUnit
 import java.io.File
+import java.io.IOException
 
+val okHttpClient = OkHttpClient()
 val terminal = Terminal()
 var version = "v3.0.0.4.od"
 var username = System.getProperty("user.name")
@@ -23,15 +27,19 @@ var osname = System.getProperty("os.name")
 var computername = InetAddress.getLocalHost().hostName
 var dir = System.getProperty("user.dir")
 var homedir = System.getProperty("user.home")
-var cursor  = ""
+var cursor  = "${cyan("($username @ $computername : $dir )")} ${green("~>")} "
 var shellCommands = arrayOf("fs", "exit", "cd", "fpm")
 val freeshellLinuxPath = File("/home/$username/.local/share/.freeshell")
+val fpmLinuxPath = File("/home/$username/.local/share/.freeshell/fpm")
 val freeshellWindowsPath = File("C:/Users/$username/AppData/Roaming/.freeshell")
+val fpmWindowsPath = File("C:/Users/$username/AppData/Roaming/.freeshell")
+
 val strfshlxPath = "/home/$username/.local/share/.freeshell"
 val strfshwinPath = "C:/Users/$username/AppData/Roaming/.freeshell"
 val fshLinuxExists = freeshellLinuxPath.exists()
 val fshWindowsExists = freeshellWindowsPath.exists()
 var fpmIsEnabled = true;
+
 fun main(args: Array<String>){
     if (osname.startsWith("Windows")){
         printConsoleWarning("Windows commands are not well supported")
@@ -46,32 +54,43 @@ fun main(args: Array<String>){
 
     }
 
-    if (!fshLinuxExists || !freeshellLinuxPath.isDirectory && !fshWindowsExists || !freeshellWindowsPath.isDirectory) {
+    if (!fshLinuxExists && !freeshellLinuxPath.isDirectory && !fshWindowsExists && !freeshellWindowsPath.isDirectory) {
         println("${yellow("!:")} Freeshell Config not found. Continuing with setup.")
         Thread.sleep(2000)
         terminal.println("Freeshell Setup", Whitespace.NORMAL, TextAlign.CENTER, OverflowWrap.NORMAL)
         println("---")
         if (osname.startsWith("Windows")) {
-            freeshellWindowsPath.mkdir();
+            freeshellWindowsPath.mkdir()
         } else {
             freeshellLinuxPath.mkdir()
         }
-        val response = questionSystem("Do you want to install FPM?", listOf("yes", "no"))
-        if (response == "yes") {
+        val installFPM = questionSystem("Do you want to install FPM?", listOf("yes", "no"))
+        if (installFPM == "yes") {
             terminal.println(bold("Freeshell Setup. FPM"))
-            terminal.prompt("${blue("?:")} Choose your fpm version (check https://github.com/project-novagon/fpm/releases/latest for the latest release)", choices= listOf("latest", "choose"))
-            printConsoleInfo("Installing FPM...", false)
-            /*
+            if (osname.startsWith("Windows")) fpmWindowsPath.mkdir() else fpmLinuxPath.mkdir()
+            printConsoleWarning("Automatically installing latest FPM verison...")
 
-                Fuel.get("https://github.com/project-novagon/fpm/releases/download/v$version/fpm.py")
-                    .response { request, response, result ->
-                        val (bytes, error) = result
-                        if (bytes != null) {
-                            File("path/goes/here").writeBytes(bytes)
-                        }
-                    }
-            */
-            printConsoleInfo("FPM Installed!", true)
+            val request = Request.Builder()
+                .url("https://api.github.com/repos/project-novagon/fpm/releases/latest")
+                .build()
+
+            okHttpClient.newCall(request).execute().use { response ->
+                if (!response.isSuccessful) throw IOException("Unexpected code $response")
+
+                val latestRelease = response.body!!.string()
+                val tagName = latestRelease.substringAfter("\"tag_name\":\"").substringBefore("\",\"target_commitish\"")
+                println(tagName)
+
+                printConsoleInfo("Installing FPM...", false)
+                val fpmRequest = Request.Builder()
+                    .url("https://github.com/project-novagon/fpm/releases/download/$tagName/fpm.py")
+                    .build()
+
+                val fpmfile = okHttpClient.newCall(fpmRequest).execute()
+                File("${if (osname.startsWith("Windows")) "$fpmWindowsPath/fpm.py" else "$fpmLinuxPath/fpm.py"}").writeBytes(fpmfile.body!!.bytes())
+                printConsoleInfo("FPM Installed!", true)
+            }
+
 
         }
         else {
